# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Productivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

/**
 * @nullable
 */
export type ProfileProfession = typeof ProfileProfession[keyof typeof ProfileProfession] | null;


export const ProfileProfession = {
  Student: 'Student',
  Employee: 'Employee',
  Athlete: 'Athlete',
  Freelancer: 'Freelancer',
  Custom: 'Custom',
} as const;

export interface Profile {
  id: number;
  userId: number;
  name: string;
  email: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  profession?: ProfileProfession;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  goals?: string | null;
  /** @nullable */
  wakeUpTime?: string | null;
  /** @nullable */
  sleepTarget?: number | null;
}

/**
 * @nullable
 */
export type ProfileUpdateProfession = typeof ProfileUpdateProfession[keyof typeof ProfileUpdateProfession] | null;


export const ProfileUpdateProfession = {
  Student: 'Student',
  Employee: 'Employee',
  Athlete: 'Athlete',
  Freelancer: 'Freelancer',
  Custom: 'Custom',
} as const;

export interface ProfileUpdate {
  name?: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  profession?: ProfileUpdateProfession;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  goals?: string | null;
  /** @nullable */
  wakeUpTime?: string | null;
  /** @nullable */
  sleepTarget?: number | null;
}

export interface Activity {
  id: number;
  userId: number;
  name: string;
  category: string;
  isProductive: boolean;
  /** @nullable */
  targetHours?: number | null;
  /** @nullable */
  color?: string | null;
}

export interface ActivityInput {
  /** @minLength 1 */
  name: string;
  category: string;
  isProductive: boolean;
  /** @nullable */
  targetHours?: number | null;
  /** @nullable */
  color?: string | null;
}

export interface ActivityUpdate {
  name?: string;
  category?: string;
  isProductive?: boolean;
  /** @nullable */
  targetHours?: number | null;
  /** @nullable */
  color?: string | null;
}

export interface LogEntry {
  id: number;
  userId: number;
  activityId: number;
  activityName?: string;
  date: string;
  completed: boolean;
  /** @nullable */
  hoursSpent?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  productivityPct?: number | null;
}

export interface LogEntryInput {
  activityId: number;
  completed: boolean;
  /** @nullable */
  hoursSpent?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  productivityPct?: number | null;
}

export interface DashboardSummary {
  todayScore: number;
  currentStreak: number;
  longestStreak?: number;
  weeklyImprovement: number;
  completedToday: number;
  totalToday: number;
  productiveHoursToday: number;
  totalHoursToday?: number;
  /** @nullable */
  bmi: number | null;
  /** @nullable */
  bmiCategory?: string | null;
  badges?: string[];
}

export interface DayStats {
  date: string;
  dayLabel: string;
  score: number;
  completedActivities: number;
  totalActivities: number;
  productiveHours: number;
}

export interface WeeklyAnalytics {
  days: DayStats[];
  weeklyAvgScore?: number;
  previousWeekAvgScore?: number;
}

export interface WeekStats {
  weekLabel: string;
  avgScore: number;
  activeDays: number;
}

export interface MonthlyAnalytics {
  weeks: WeekStats[];
  monthlyAvgScore?: number;
  consistencyPct?: number;
}

export interface ActivityBreakdownItem {
  name: string;
  hours: number;
  pct: number;
  isProductive: boolean;
  color: string;
}

export interface ActivityBreakdown {
  items: ActivityBreakdownItem[];
  productiveHours: number;
  nonProductiveHours: number;
  productivePct: number;
}

export type TipPriority = typeof TipPriority[keyof typeof TipPriority];


export const TipPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Tip {
  category: string;
  message: string;
  priority: TipPriority;
}

export interface Recommendations {
  strengths: string[];
  weaknesses: string[];
  tips: Tip[];
  weeklyInsight: string;
  productivityAnalysis?: string;
}

export interface HistoryEntry {
  date: string;
  score: number;
  completedActivities: number;
  totalActivities: number;
  productiveHours: number;
  entries?: LogEntry[];
}

export type GetHistoryParams = {
startDate?: string;
endDate?: string;
groupBy?: GetHistoryGroupBy;
};

export type GetHistoryGroupBy = typeof GetHistoryGroupBy[keyof typeof GetHistoryGroupBy];


export const GetHistoryGroupBy = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;