# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Productivity Intelligence Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: profile
    description: User profile operations
  - name: activities
    description: Activity management
  - name: logs
    description: Daily activity logs
  - name: analytics
    description: Productivity analytics
  - name: recommendations
    description: AI-like recommendations
  - name: history
    description: Historical data
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List user activities
      responses:
        "200":
          description: Activities list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create a new activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

  /activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /logs/today:
    get:
      operationId: getTodayLog
      tags: [logs]
      summary: Get today's activity log
      responses:
        "200":
          description: Today's log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"

  /logs:
    post:
      operationId: upsertLogEntry
      tags: [logs]
      summary: Create or update a log entry for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogEntryInput"
      responses:
        "200":
          description: Upserted log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogEntry"

  /logs/{id}:
    delete:
      operationId: deleteLogEntry
      tags: [logs]
      summary: Delete a log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /analytics/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [analytics]
      summary: Get dashboard summary with productivity score, streaks, and today stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /analytics/weekly:
    get:
      operationId: getWeeklyAnalytics
      tags: [analytics]
      summary: Get weekly productivity data for charts
      responses:
        "200":
          description: Weekly analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyAnalytics"

  /analytics/monthly:
    get:
      operationId: getMonthlyAnalytics
      tags: [analytics]
      summary: Get monthly consistency and trend data
      responses:
        "200":
          description: Monthly analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyAnalytics"

  /analytics/activity-breakdown:
    get:
      operationId: getActivityBreakdown
      tags: [analytics]
      summary: Get activity distribution (productive vs non-productive)
      responses:
        "200":
          description: Activity breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityBreakdown"

  /recommendations:
    get:
      operationId: getRecommendations
      tags: [recommendations]
      summary: Get AI-like personalized recommendations based on behavior
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendations"

  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get historical logs with date range filter
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: Historical data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string

    Profile:
      type: object
      required: [id, userId, name, email]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        profession:
          type: ["string", "null"]
          enum: [Student, Employee, Athlete, Freelancer, Custom, null]
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        goals:
          type: ["string", "null"]
        wakeUpTime:
          type: ["string", "null"]
        sleepTarget:
          type: ["number", "null"]

    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        profession:
          type: ["string", "null"]
          enum: [Student, Employee, Athlete, Freelancer, Custom, null]
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        goals:
          type: ["string", "null"]
        wakeUpTime:
          type: ["string", "null"]
        sleepTarget:
          type: ["number", "null"]

    Activity:
      type: object
      required: [id, userId, name, category, isProductive]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        category:
          type: string
        isProductive:
          type: boolean
        targetHours:
          type: ["number", "null"]
        color:
          type: ["string", "null"]

    ActivityInput:
      type: object
      required: [name, category, isProductive]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        isProductive:
          type: boolean
        targetHours:
          type: ["number", "null"]
        color:
          type: ["string", "null"]

    ActivityUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        isProductive:
          type: boolean
        targetHours:
          type: ["number", "null"]
        color:
          type: ["string", "null"]

    LogEntry:
      type: object
      required: [id, userId, activityId, date, completed]
      properties:
        id:
          type: integer
        userId:
          type: integer
        activityId:
          type: integer
        activityName:
          type: string
        date:
          type: string
          format: date
        completed:
          type: boolean
        hoursSpent:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        productivityPct:
          type: ["number", "null"]

    LogEntryInput:
      type: object
      required: [activityId, completed]
      properties:
        activityId:
          type: integer
        completed:
          type: boolean
        hoursSpent:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        productivityPct:
          type: ["number", "null"]

    DashboardSummary:
      type: object
      required: [todayScore, currentStreak, weeklyImprovement, completedToday, totalToday, productiveHoursToday, bmi]
      properties:
        todayScore:
          type: number
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        weeklyImprovement:
          type: number
        completedToday:
          type: integer
        totalToday:
          type: integer
        productiveHoursToday:
          type: number
        totalHoursToday:
          type: number
        bmi:
          type: ["number", "null"]
        bmiCategory:
          type: ["string", "null"]
        badges:
          type: array
          items:
            type: string

    WeeklyAnalytics:
      type: object
      required: [days]
      properties:
        days:
          type: array
          items:
            $ref: "#/components/schemas/DayStats"
        weeklyAvgScore:
          type: number
        previousWeekAvgScore:
          type: number

    DayStats:
      type: object
      required: [date, dayLabel, score, completedActivities, totalActivities, productiveHours]
      properties:
        date:
          type: string
        dayLabel:
          type: string
        score:
          type: number
        completedActivities:
          type: integer
        totalActivities:
          type: integer
        productiveHours:
          type: number

    MonthlyAnalytics:
      type: object
      required: [weeks]
      properties:
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/WeekStats"
        monthlyAvgScore:
          type: number
        consistencyPct:
          type: number

    WeekStats:
      type: object
      required: [weekLabel, avgScore, activeDays]
      properties:
        weekLabel:
          type: string
        avgScore:
          type: number
        activeDays:
          type: integer

    ActivityBreakdown:
      type: object
      required: [items, productiveHours, nonProductiveHours, productivePct]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ActivityBreakdownItem"
        productiveHours:
          type: number
        nonProductiveHours:
          type: number
        productivePct:
          type: number

    ActivityBreakdownItem:
      type: object
      required: [name, hours, pct, isProductive, color]
      properties:
        name:
          type: string
        hours:
          type: number
        pct:
          type: number
        isProductive:
          type: boolean
        color:
          type: string

    Recommendations:
      type: object
      required: [strengths, weaknesses, tips, weeklyInsight]
      properties:
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        tips:
          type: array
          items:
            $ref: "#/components/schemas/Tip"
        weeklyInsight:
          type: string
        productivityAnalysis:
          type: string

    Tip:
      type: object
      required: [category, message, priority]
      properties:
        category:
          type: string
        message:
          type: string
        priority:
          type: string
          enum: [high, medium, low]

    HistoryEntry:
      type: object
      required: [date, score, completedActivities, totalActivities, productiveHours]
      properties:
        date:
          type: string
        score:
          type: number
        completedActivities:
          type: integer
        totalActivities:
          type: integer
        productiveHours:
          type: number
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LogEntry"