# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Productivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityInput {
  /** @minLength 1 */
  name: string;
  category: string;
  isProductive: boolean;
  /** @nullable */
  targetHours?: number | null;
  /** @nullable */
  color?: string | null;
}