# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Productivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  todayScore: number;
  currentStreak: number;
  longestStreak?: number;
  weeklyImprovement: number;
  completedToday: number;
  totalToday: number;
  productiveHoursToday: number;
  totalHoursToday?: number;
  /** @nullable */
  bmi: number | null;
  /** @nullable */
  bmiCategory?: string | null;
  badges?: string[];
}