# lib/api-zod/src/generated/types/dayStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Productivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DayStats {
  date: string;
  dayLabel: string;
  score: number;
  completedActivities: number;
  totalActivities: number;
  productiveHours: number;
}