# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Productivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LogEntry } from './logEntry';

export interface HistoryEntry {
  date: string;
  score: number;
  completedActivities: number;
  totalActivities: number;
  productiveHours: number;
  entries?: LogEntry[];
}