# lib/api-zod/src/generated/types/logEntryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Productivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LogEntryInput {
  activityId: number;
  completed: boolean;
  /** @nullable */
  hoursSpent?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  productivityPct?: number | null;
}