# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Productivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileProfession } from './profileProfession';

export interface Profile {
  id: number;
  userId: number;
  name: string;
  email: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  profession?: ProfileProfession;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  goals?: string | null;
  /** @nullable */
  wakeUpTime?: string | null;
  /** @nullable */
  sleepTarget?: number | null;
}