# lib/api-zod/src/generated/types/weekStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Productivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface WeekStats {
  weekLabel: string;
  avgScore: number;
  activeDays: number;
}